move tick mark content into its own folder
closes #319
