+++ projects/ngx-d3timeline/src/lib/axis/axis.selectors.ts
@@ -5,8 +5,8 @@ import {
 } from '../options.selectors';
 import { getAxis } from './axis-utils';
 import { selectBandScale, selectTimeScale } from '../store/timeline-selectors';
-import { getResourceAxisTickMarkRenderer } from './resources-axis/resource-axis-tick-mark-renderer';
-import { getTimeAxisTickMarkRenderer } from './time-axis/time-axis-tick-mark-renderer';
+import { getResourceAxisTickMarkRenderer } from '../tick-mark/resource-axis-tick-mark-renderer';
+import { getTimeAxisTickMarkRenderer } from '../tick-mark/time-axis-tick-mark-renderer';
 import { selectViewTopLeft } from '../view/view.selectors';
 
 export const selectResourceAxis = createSelector(
