+++ projects/ngx-d3timeline/src/lib/tick-mark/axis-tick-mark.component.ts
