+++ projects/ngx-d3timeline/src/lib/tick-mark/tick-mark.ts
@@ -1,4 +1,4 @@
-import { Line } from './line';
+import { Line } from '../axis/line';
 import { Point } from '../point';
 
 export interface TickMark {
